refactor(tps-monitoring): update author and license in package.json, enhance README content
- Changed author name to "Simple TPS Monitor Team" and updated license to "UNLICENSED" in package.json
- Translated README content to English for broader accessibility
- Improved clarity and consistency in usage instructions and features section
